requirements.txt
Flask>=2.2.0
folium==0.12.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
